Commit script updated

+++ src/opt/boilercontrol/gitcreateinitcopyandcommit.sh
@@ -3,15 +3,19 @@
 # Run like that:
 # sudo ./gitcreateinitcopyandcommit.sh "<YOUR-COMMIT-COMMENT-HERE>"
 
-rm -rf hestia-dev;
-mkdir hestia-dev;
-cd hestia-dev;
-git init;
-git remote add origin https://github.com/gulliverrr/hestia-dev.git;
-git pull origin master;
-sudo cp -r /opt/boilercontrol/* src/opt/boilercontrol/;
-sudo find . -name "*.class" -exec rm -rf {} \;
-sudo git add src;
-sudo git commit -m "$1";
-git push origin master;
-
+if [ $# -eq 0 ]
+  then
+    echo "Usage: sudo ./gitcreateinitcopyandcommit.sh \"<YOUR-COMMIT-COMMENT-HERE>\""
+  else
+    rm -rf hestia-dev;
+    mkdir hestia-dev;
+    cd hestia-dev;
+    git init;
+    git remote add origin https://github.com/gulliverrr/hestia-dev.git;
+    git pull origin master;
+    sudo cp -r /opt/boilercontrol/* src/opt/boilercontrol/;
+    sudo find . -name "*.class" -exec rm -rf {} \;
+    sudo git add src;
+    sudo git commit -m "$1";
+    git push origin master;
+fi
